pyproject.toml
[project]
name = "photoblog"
version = "0.1.0"
description = "A Django project"
authors = [
  { name = "Your Name", email = "you@example.com" },
]
requires-python = "==3.14.*"
readme = "README.md"
license = { text = "MIT" }

dependencies = [
  "aiohttp>=3.13.3",
  "django-allauth[socialaccount]>=64.2.0",
  "django-anymail>=11.1",
  "django-health-check>=3.18.3",
  "django-htmx>=1.19.0",
  "django-http-compression[brotli]>=1.2.0",
  "django-linear-migrations>=2.13.0",
  "django-permissions-policy>=4.21.0",
  "django-redis>=5.4.0",
  "django-tailwind-cli>=2.16.0",
  "django-tasks-db>=0.12.0",
  "django-widget-tweaks>=1.5.1",
  "django>=6.0a1,<6.1",
  "environs[django]>=11.0.0",
  "gunicorn>=23.0.0",
  "heroicons[django]>=2.8.0",
  "opentelemetry-api>=1.28.0",
  "opentelemetry-exporter-otlp-proto-http>=1.28.0",
  "opentelemetry-exporter-otlp>=1.28.0",
  "opentelemetry-instrumentation-django>=0.49b0",
  "opentelemetry-instrumentation-psycopg>=0.60b1",
  "opentelemetry-instrumentation-redis>=0.49b0",
  "opentelemetry-instrumentation-requests>=0.60b1",
  "opentelemetry-sdk>=1.28.0",
  "psutil>=6.0.0",
  "psycopg[binary,pool]>=3.2.1",
  "redis>=5.0.8",
  "sentry-sdk>=2.13.0",
  "uvicorn[standard]>=0.41.0",
  "whitenoise[brotli]>=6.11.0",
  "django-storages[s3]>=1.14.4",
  "sorl-thumbnail>=13.0.0",
  "pillow>=12.1.1",
]
[dependency-groups]
dev = [
  "aioresponses>=0.7.8",
  "basedpyright>=1.34.0",
  "django-stubs-ext>=5.0.4",
  "django-stubs>=5.0.4",
  "coverage>=7.6.1",
  "django-browser-reload>=1.15.0",
  "django-debug-toolbar>=4.4.6",
  "django-watchfiles>=1.4.0",
  "factory-boy>=3.3.1",
  "faker>=33.0.0",
  "pytest-asyncio>=1.3.0",
  "pytest-cov>=5.0.0",
  "pytest-django>=4.8.0",
  "pytest-env>=1.1.3",
  "pytest-icdiff>=0.9",
  "pytest-mock>=3.14.0",
  "pytest-playwright>=0.5.0",
  "pytest-randomly>=3.15.0",
  "pytest-watcher>=0.4.3",
  "pytest-xdist>=3.6.1",
  "pytest>=8.3.2",
  "ruff>=0.15.2",
  "djlint>=1.36.4",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
asyncio_mode = "auto"
addopts = [
  "-v",
  "-x",
  "-p no:warnings",
  "--ff",
  "--cov",
  "--reuse-db",
  "--no-cov-on-fail",
  # Uncomment when the project is mature and you want to enforce full coverage:
  # "--cov-fail-under=100",
  "--cov-report=term-missing",
  "--tb=short",
  "-m",
  "not e2e",
]
markers = ["e2e: end-to-end browser tests with Playwright"]

testpaths = ["photoblog", "templates"]
env = [
  "COVERAGE_CORE=sysmon",
  "DJANGO_ALLOW_ASYNC_UNSAFE=true",
  "USE_CONNECTION_POOL=false",
  "USE_COLLECTSTATIC=false",
  "USE_HTTPS=false",
  "USE_X_FORWARDED_HOST=false",
]

[tool.pytest-watcher]
now = false
clear = true
delay = 0.2
runner = "pytest"
patterns = ["*.py", "*.html"]

[tool.coverage.run]
source = ["photoblog"]
omit = ["**/migrations/*", "*/tests/*"]

[tool.djlint]
profile = "django"
ignore = "H016,H030,H031,H017,H021,T003"
custom_blocks = "cache,partialdef,fragment"

[tool.ruff]
target-version = "py314"


lint.select = [
  "B",
  "C4",
  "C90",
  "D10",
  "DJ",
  "E",
  "ERA",
  "F",
  "FA",
  "FBT",
  "G",
  "I001",
  "INP",
  "N",
  "PGH",
  "PIE",
  "PL",
  "PT",
  "PTH",
  "RET",
  "RSE",
  "RUF",
  "S",
  "SIM",
  "T20",
  "TCH",
  "TID",
  "W",
]
lint.ignore = [
  "D100",
  "D104",
  "D106",
  "D107",
  "D203",
  "D204",
  "D213",
  "D215",
  "D400",
  "D401",
  "D404",
  "D406",
  "D407",
  "D408",
  "D409",
  "D413",
  "DJ008",
  "E501",
  "E701",
  "PLR0913",
  "PT019",
  "Q000",
  "S101",
]

[tool.ruff.lint.per-file-ignores]
"**/apps.py" = ["D101", "D102", "D103", "D105", "PL", "RUF"]
"**/conftest.py" = ["D101", "D102", "D103", "D105", "PL", "RUF"]
"**/migrations/**.py" = [
  "D101",
  "D102",
  "D103",
  "D105",
  "E501",
  "N",
  "PL",
  "RUF",
]
"**/tests/**.py" = ["D101", "D102", "D103", "D105", "PL", "RUF"]

[tool.ruff.lint.isort]
known-first-party = ["photoblog"]


[tool.django-stubs]
django_settings_module = "config.settings"

[tool.pyright]
include = ["photoblog"]
exclude = ["**/migrations/*.py", "**/tests/**/*.py"]
typeCheckingMode = "basic"
